move the details to a new folder and change the routes
